[docs] clarify what "it" refers to (#4148)
* Update 03-loading.md

Edits ambiguous pronoun to make it clear that 'load' could run on either the client or the server.  I have seen a couple of tutorials recently with content creators stating the putting things like client secrets in this function within <script context="module"> is okay because it runs on the server.  If my understanding is correct, this is not safe and could possibly be made a little more clear.

* Update documentation/docs/03-loading.md

Co-authored-by: Ben McCann <322311+benmccann@users.noreply.github.com>